Check sframe 0 and 1 first, a bit faster
I think the biggest benefit there is avoiding an extra branch and
multiply, of sframes 0 and 1 only 1 requires a multiply, while both
sframes 2 and 3 do. This does result in more "half matches" (which
require the high half of the checksum to be computed), but they're
still rare enough (about 1 in 64k) that it's worthwhile. It's about
30% faster in my tests.

This change also runs through all hwords in a single invocation, with
some stride. It would probably be better to have an external work
scheduler manage the bootcsumr instances, but this is better for my use
cases.